requirements.txt
darts==0.24.0
faiss-cpu==1.7.4
fastapi==0.95.1
langchain==0.0.154
openai==0.27.5
pandas==1.5.3
pdfminer.six==20221105
python-dotenv==1.0.0
redis==4.5.4
requests==2.29.0
tiktoken==0.4.0
uvicorn==0.22.0
rocketry==2.5.1